+++ src/Angular2Apollo.ts
@@ -1,12 +1,10 @@
 import { Provider, provide, OpaqueToken, Injectable, Inject } from '@angular/core';
+import { rxify } from 'apollo-client-rxjs';
+import { ApolloQueryResult } from 'apollo-client';
 
 import { ApolloQueryObservable } from './ApolloQueryObservable';
-import { ObservableQueryRef } from './utils/ObservableQuery';
-import { observeVariables } from './utils/observeVariables';
 
 import ApolloClient from 'apollo-client';
-import assign = require('lodash.assign');
-import omit = require('lodash.omit');
 
 import 'rxjs/add/operator/switchMap';
 
@@ -23,27 +21,8 @@ export class Angular2Apollo {
     @Inject(angularApolloClient) private client: any
   ) {}
 
-  public watchQuery(options): ApolloQueryObservable<any> {
-    const apolloRef = new ObservableQueryRef();
-    if (typeof options.variables === 'object') {
-      const varObs = observeVariables(options.variables);
-
-      return new ApolloQueryObservable(apolloRef, subscriber => {
-        const sub = varObs.switchMap(newVariables => {
-          const cleanOptions = omit(options, 'variables');
-          const newOptions = assign(cleanOptions, { variables: newVariables });
-
-          apolloRef.apollo = this.client.watchQuery(newOptions);
-
-          return apolloRef.apollo;
-        }).subscribe(subscriber);
-
-        return () => sub.unsubscribe();
-      });
-    }
-
-    apolloRef.apollo = this.client.watchQuery(options);
-    return new ApolloQueryObservable(apolloRef);
+  public watchQuery(options): ApolloQueryObservable<ApolloQueryResult> {
+    return new ApolloQueryObservable(rxify(this.client.watchQuery)(options));
   }
 
   public query(options) {